requirements.txt
torch
jj-wingman
numpy
opencv-python
matplotlib
git+https://github.com/dronekit/dronekit-python
pyserial
